requirements.txt
tensorflow-gpu==2.0.0-beta1